+++ lib/pages/detectionPages/cornTypeDetectionPage.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/cupertino.dart';
+
+class CornTypeDetectionPage extends StatefulWidget {
+  const CornTypeDetectionPage({super.key});
+
+  @override
+  State<CornTypeDetectionPage> createState() => _CornTypeDetectionPageState();
+}
+
+class _CornTypeDetectionPageState extends State<CornTypeDetectionPage> {
+  @override
+  Widget build(BuildContext context) {
+    return const Placeholder();
+  }
+}
