+++ lib/pages/detectionPages/liveDetectionPage.dart
@@ -6,9 +6,7 @@ class LiveDetectionPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-        title: const Text('Live Detection'),
-      ),
+      
       body: const Center(
         child: Text(
           'This is the Live Detection Page',
