+++ lib/pages/homeDetection.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:seedscan2/pages/detectionPages/imageDetectionPage.dart';
 import 'package:seedscan2/pages/detectionPages/liveDetectionPage.dart';
+import 'package:seedscan2/pages/detectionPages/imageDetectionPage.dart';
 
 class HomeDetection extends StatefulWidget {
   const HomeDetection({super.key});
@@ -14,33 +14,79 @@ class HomeDetectionState extends State<HomeDetection> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: const Text('Home Detection'),
-        centerTitle: true,
       ),
       body: Center(
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
           children: [
-            ElevatedButton(
-              onPressed: () {
-                // Navigate to the LiveDetectionPage
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(builder: (context) => const LiveDetectionPage()),
-                );
-              },
-              child: const Text('Detect Live'),
+            SizedBox(
+              width: MediaQuery.of(context).size.width * 0.9, // 90% of the screen width
+              child: Card(
+                elevation: 3,
+                margin: const EdgeInsets.symmetric(vertical: 10),
+                child: Padding(
+                  padding: const EdgeInsets.all(20.0),
+                  child: Column(
+                    children: [
+                      const Text(
+                        'Live Detection',
+                        style: TextStyle(fontSize: 18, fontWeight: FontWeight.bold),
+                      ),
+                      const SizedBox(height: 10),
+                      ElevatedButton(
+                        onPressed: () {
+                          Navigator.push(
+                            context,
+                            MaterialPageRoute(builder: (context) => const LiveDetectionPage()),
+                          );
+                        },
+                        style: ElevatedButton.styleFrom(
+                          backgroundColor: Colors.blue, // Button background color
+                          foregroundColor: Colors.white, // Button text color
+                          padding: const EdgeInsets.symmetric(vertical: 15, horizontal: 30),
+                          textStyle: const TextStyle(fontSize: 16, fontWeight: FontWeight.bold),
+                        ),
+                        child: const Text('Detect Live'),
+                      ),
+                    ],
+                  ),
+                ),
+              ),
             ),
-            const SizedBox(height: 20), // Add spacing between buttons
-            ElevatedButton(
-              onPressed: () {
-                // Navigate to the ImageDetectionPage
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(builder: (context) => const ImageDetectionPage()),
-                );
-              },
-              child: const Text('Detect Image'),
+            const SizedBox(height: 40), // Gap space between cards
+            SizedBox(
+              width: MediaQuery.of(context).size.width * 0.9, // 90% of the screen width
+              child: Card(
+                elevation: 3,
+                margin: const EdgeInsets.symmetric(vertical: 10),
+                child: Padding(
+                  padding: const EdgeInsets.all(20.0),
+                  child: Column(
+                    children: [
+                      const Text(
+                        'Image Detection',
+                        style: TextStyle(fontSize: 18, fontWeight: FontWeight.bold),
+                      ),
+                      const SizedBox(height: 10),
+                      ElevatedButton(
+                        onPressed: () {
+                          Navigator.push(
+                            context,
+                            MaterialPageRoute(builder: (context) => const ImageDetectionPage()),
+                          );
+                        },
+                        style: ElevatedButton.styleFrom(
+                          backgroundColor: Colors.green, // Button background color
+                          foregroundColor: Colors.white, // Button text color
+                          padding: const EdgeInsets.symmetric(vertical: 15, horizontal: 30),
+                          textStyle: const TextStyle(fontSize: 16, fontWeight: FontWeight.bold),
+                        ),
+                        child: const Text('Detect Image'),
+                      ),
+                    ],
+                  ),
+                ),
+              ),
             ),
           ],
         ),
